# <format>: Compile-time string size estimation