--- a/tests/std/tests/P0645R10_text_formatting_formatting/env.lst
+++ b/tests/std/tests/P0645R10_text_formatting_formatting/env.lst
@@ -2,3 +2,6 @@
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
 RUNALL_INCLUDE ..\concepts_latest_matrix.lst
+RUNALL_CROSSLIST
+PM_CL=""
+PM_CL="/utf-8"