… stl/inc/xstring
… tests/std/tests/P0645R10_text_formatting_formatting/test.cpp