Improves string size estimation in `format` by using `_Basic_format_string`'s parsing machinery. It takes into account things like argument use count and the specified width and precision. Also allows `formatted_size` to completely skip the formatting part for simple cases like concatenation of a bunch of strings.

This change might be ABI breaking so hopefully it can make it before `<format>`'s ABI is freezed.