@barcharcraz 2022-02-03
We talked about this PR in our weekly meeting, and I think the next steps are to write a small benchmark and test out a simpler version with only the fallback path. We consider this too complex to merge without a benchmark showing very clear improvement.