@AdamBucior 2022-02-04
With this simple benchmark:
```
#include <string>
#include <format>
#include <chrono>
#include <iostream>

using namespace std;
using namespace chrono;

constexpr size_t Iterations = 10000000;

int main() {
    steady_clock::time_point start, end;
    start = steady_clock::now();

    for (size_t i = 0; i < Iterations; ++i) {
        (void) format("{} {} {} {} {} {}", "let's", "concatenate", "a", "bunch", "of", "strings");
    }

    end = steady_clock::now();

    cout << end - start << endl;

    start = steady_clock::now();

    for (size_t i = 0; i < Iterations; ++i) {
        (void) format("{0} {0} {0}", "repeat this sentence three times");
    }

    end = steady_clock::now();

    cout << end - start << endl;

    system("pause");
}
```

I got these results (ran for 4 times each):

```
With PR:

4473119900ns
5105820100ns

4532696100ns
5093983400ns

4463521400ns
5120882300ns

4421421200ns
5108412600ns

Average:
4472689650ns
5107274600ns

Without PR:

4750032000ns
5193818700ns

4790922400ns
5249655700ns

5018952300ns
5281830700ns

4815681900ns
5280669900ns

Average:
4843897150ns ~8% more time than average with PR
5251493750ns ~3% more time than average with PR
```

So with simple strings there is some improvement. Obviously the result will differ depending on arguments used (of which there is a wide variety of possibilities). I don't think there is really a case in which this PR would perform worse.