@barcharcraz 2022-02-08
What do you get running that with clang-cl instead? (to do just the fallback path)