@AdamBucior 2022-02-08
With clang-cl:
```
With PR:

3826073500ns
4233170700ns

3630934800ns
4309987800ns

3822146900ns
4210166300ns

3612824000ns
4434587900ns

Average:

3722994800ns
4296978175ns

Without PR:

3760814900ns
4333771500ns

3692610400ns
4379408800ns

3719316200ns
4294400100ns

3744752400ns
4294775400ns

Average:

3729373475ns ~0.1% more time than average with PR (no real difference)
4325588950ns ~0.6% more time than average with PR (no real difference)
```
Unsurprisingly no real difference. The fallback is basically the same thing that we are already doing.