@StephanTLavavej 2022-02-09
@barcharcraz says that there should be no inherent ABI impact to these changes - at most, we may need to rename an `_Ugly` internal type to `_Ugly2` (as we've done repeatedly in `shared_ptr`'s control blocks), which prevents any mix-and-match problems. Charlie would also prefer not to backport this to VS 2019 16.11.x, so we should ensure that before-and-after code can be mixed harmlessly, even if these changes get into VS 2022 17.2 Preview 3 (before the 17.2 General Availability ABI lockdown).

Charlie plans to investigate this with more benchmarking.