@barcharcraz 2022-04-01
I benchmarked this using gbench (it did around 3,000,000 iterations for each test) and found an extremely small perf difference even with the full caching (I verified I was using the correct version by adding a define in format). It seems like this pr does make things a new nanoseconds faster but not always.

In general doing three benchmark iterations on a benchmark that is so fast won't give reliable results, and this is true even if the CPU is literally executing nothing else (i.e. the benchmark is running in kernel-mode, all alone). Here are my results

```
Has cache: 0
2022-03-31T17:16:03-07:00
Running C:\Users\chbarto\source\repos\scratch\format_caching_bench\build\bench.exe
Run on (12 X 3192 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x6)
  L1 Instruction 32 KiB (x6)
  L2 Unified 256 KiB (x6)
  L3 Unified 12288 KiB (x1)
-------------------------------------------------------
Benchmark             Time             CPU   Iterations
-------------------------------------------------------
bench_concat        225 ns          225 ns      2986667
bench_insert        349 ns          353 ns      2036364
```
Has cache: 1
2022-03-31T17:16:50-07:00
Running C:\Users\chbarto\source\repos\scratch\format_caching_bench\build\bench.exe
Run on (12 X 3192 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x6)
  L1 Instruction 32 KiB (x6)
  L2 Unified 256 KiB (x6)
  L3 Unified 12288 KiB (x1)
-------------------------------------------------------
Benchmark             Time             CPU   Iterations
-------------------------------------------------------
bench_concat        227 ns          225 ns      2986667
bench_insert        328 ns          330 ns      2133333
```

code:
```c++
#include <benchmark/benchmark.h>
#include <stdio.h>
#include <format>
#ifndef _STL_FORMAT_HAS_CACHE
#define _STL_FORMAT_HAS_CACHE 0
#endif
using benchmark::State;
static void bench_concat(State &s)
{
    for (auto _ : s)
    {
        (void)std::format("{} {} {} {} {} {}", "let's", "concatenate", "a", "bunch", "of", "strings");
    }
}

static void bench_insert(State &s)
{
    for (auto _ : s)
    {
        (void)std::format("{0} {0} {0}", "repeat this sentence three times");
    }
}
BENCHMARK(bench_concat);
BENCHMARK(bench_insert);
int main(int argc, char **argv)
{
    printf("Has cache: %d\n", _STL_FORMAT_HAS_CACHE);
    ::benchmark::Initialize(&argc, argv);
    if (::benchmark::ReportUnrecognizedArguments(argc, argv))
        return 1;
    ::benchmark::RunSpecifiedBenchmarks();
    ::benchmark::Shutdown();
    return 0;
}
```