@barcharcraz 2022-04-01
and here's with just the fallback caching
```
Has cache: 2
2022-03-31T17:26:47-07:00
Running C:\Users\chbarto\source\repos\scratch\format_caching_bench\build\bench.exe
Run on (12 X 3192 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x6)
  L1 Instruction 32 KiB (x6)
  L2 Unified 256 KiB (x6)
  L3 Unified 12288 KiB (x1)
-------------------------------------------------------
Benchmark             Time             CPU   Iterations
-------------------------------------------------------
bench_concat        243 ns          246 ns      2800000
bench_insert        328 ns          329 ns      1947826
```