@StephanTLavavej 2022-04-06
Thanks @barcharcraz for the additional benchmark results.

@AdamBucior, thanks for investigating this potential optimization. After looking at the perf numbers from your benchmark and Charlie's, we talked about this at the weekly maintainer meeting and have decided that the complexity of this optimization outweighs the minimal potential for performance improvement here.